+++ EDDI/MainWindow.xaml.cs
@@ -432,7 +432,7 @@ namespace Eddi
 
         private void ShipRoleChanged(object sender, SelectionChangedEventArgs e)
         {
-            if (shipsConfiguration != null)
+            if (shipsConfiguration != null && CompanionAppService.Instance.CurrentState != CompanionAppService.State.READY)
             {
                 shipsConfiguration.ToFile();
             }
@@ -440,7 +440,7 @@ namespace Eddi
 
         private void shipYardUpdated(object sender, DataTransferEventArgs e)
         {
-            if (shipsConfiguration != null)
+            if (shipsConfiguration != null && CompanionAppService.Instance.CurrentState != CompanionAppService.State.READY)
             {
                 shipsConfiguration.ToFile();
             }